Add files via upload
Hi! So this is my first project, which a very basic login page which I had created for gdsc inductions. It consists of three buttons two of which add two input fields of user name and password and the third button links to the google sign in page. It only has frontend.